# master/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: birdman master API
  version: v1
  description: |-
    Public REST API of birdman-master: matchmaking, fleet control, deploys and observability.

    This document is GENERATED from the master's route table (master/internal/httpapi/routes.go)
    and cannot drift from the router: the router is registered from that same table.

    Coverage: paths, methods, authentication and success status codes are complete.
    Response bodies are described for every endpoint except a handful that assemble
    them dynamically — each of those says so in its own response description.
    Query parameters and request bodies are not described yet.
  license:
    name: MIT
    url: https://github.com/ufna/birdman/blob/main/LICENSE
servers:
  - url: '{master}'
    description: Your own birdman master. Every install has its own address; there is no hosted service.
    variables:
      master:
        default: http://127.0.0.1:8100
        description: Base URL of the master — the default of the deploy/ compose stack.
paths:
...
  /v1/allocate:
    post:
      operationId: allocate
      summary: Allocate a ready dedicated server directly
      description: Requires an API key with the "allocate" scope (or "admin").
      security:
        - apiKey: []
      responses:
        4XX:
          description: Client error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpapi.apiError'
        5XX:
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpapi.apiError'
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/store.Allocation'
...
  /v1/nodes:
    get:
      operationId: listNodes
      summary: List fleet nodes
      description: Requires an API key with the "readonly" scope (or "admin").
      security:
        - apiKey: []
      responses:
        4XX:
          description: Client error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpapi.apiError'
        5XX:
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpapi.apiError'
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpapi.nodesResp'
    post:
      operationId: createNode
      summary: Enroll a node into the fleet
      description: Requires an API key with the "admin" scope (or "admin").
      security:
        - apiKey: []
      responses:
        4XX:
          description: Client error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpapi.apiError'
        5XX:
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpapi.apiError'
        "201":
          description: Success. This endpoint's response body is not described yet (it is assembled dynamically); see docs/specs/master.md §6.
  /v1/nodes/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    patch:
      operationId: setNodeEnv
      summary: Move a node to another environment
      description: Requires an API key with the "admin" scope (or "admin").
      security:
        - apiKey: []
      responses:
        4XX:
          description: Client error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpapi.apiError'
        5XX:
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpapi.apiError'
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpapi.nodeResp'
...
  /v1/openapi.yaml:
    get:
      operationId: openAPI
      summary: This API's own OpenAPI 3.1 contract, generated from the route table
      description: 'Public by design: no API key required.'
      security: []
      responses:
        4XX:
          description: Client error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpapi.apiError'
        5XX:
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpapi.apiError'
        "200":
          description: Success. Body is not described by this contract (an upstream proxy response, an event stream, a non-JSON payload or no body at all).
...
components:
  schemas:
...
    httpapi.apiError:
      type: object
      properties:
        detail:
          type: string
        error:
          type: string
      required:
        - detail
        - error
...
    httpapi.nodeResp:
      type: object
      properties:
        node:
          $ref: '#/components/schemas/store.Node'
      required:
        - node
    httpapi.nodesResp:
      type: object
      properties:
        nodes:
          type:
            - array
            - "null"
          items:
            $ref: '#/components/schemas/store.Node'
      required:
        - nodes
...
    store.Allocation:
      type: object
      properties:
        host:
          type: string
        port:
          type: integer
        server_id:
          type: string
      required:
        - host
        - port
        - server_id
...
    store.Node:
      type: object
      properties:
        agent_version:
          type: string
        capacity_slots:
          type: integer
        cert_not_after:
          type:
            - string
            - "null"
          format: date-time
        cert_serial:
          type:
            - string
            - "null"
        created_at:
          type: string
          format: date-time
        enrolled_at:
          type:
            - string
            - "null"
          format: date-time
        env:
          type: string
        hostname:
          type: string
        id:
          type: string
        labels:
          type:
            - object
            - "null"
          additionalProperties:
            description: any JSON value
        last_heartbeat_at:
          type:
            - string
            - "null"
          format: date-time
        project:
          type: string
        project_id:
          type: string
        public_ip:
          type: string
        region:
          type: string
        state:
          type: string
      required:
        - agent_version
        - capacity_slots
        - created_at
        - env
        - hostname
        - id
        - labels
        - project
        - project_id
        - public_ip
        - region
        - state
...
  securitySchemes:
    apiKey:
      type: http
      scheme: bearer
      description: 'birdman API key (bmk_…) in `Authorization: Bearer`. A key carries scopes (admin, deploy, matchmaking, allocate, readonly) and may be bound to one (project, environment) pair, outside of which it can reach nothing.'
    sessionCookie:
      type: apiKey
      in: cookie
      name: birdman_session
      description: Browser session cookie issued by POST /v1/session; the panel uses it instead of carrying an API key in every request.